fix activity test

+++ worker/pkg/workflows/ee/piidetect/workflows/table/activities/activities_test.go
@@ -184,7 +184,7 @@ func Test_DetectPiiLLM_Success(t *testing.T) {
 	report, ok := res.PiiColumns["email"]
 	assert.True(t, ok)
 	assert.Equal(t, PiiCategoryContact, report.Category)
-	assert.Equal(t, 0.95, report.Confidence)
+	assert.Equal(t, float32(0.95), report.Confidence)
 }
 
 func Test_SaveTablePiiDetectReport_Success(t *testing.T) {
